Create ApplicationQueue.Java
Array-based heap implementation of a priority queue containing Applications. Guarantees the min-heap invariant, so that the Application at the root should have the lowest score, and children always have a higher or equal score as their parent. The root of a non-empty queue is always at index 0 of this array-heap.